First pass of runable BrainRun.py this also includes an optimization to track the history connection time instead of updating it. This improved the processing time by over 100%.  Added updated documentation to the Globals.py Fixed a number of bugs durring the brainRun testing.
In the workspace I also added tools for profiling.

We are using -module cProfile -o tmp.profile when running the
code to profile.
Use snakeviz tmp.profile to review the results.

One optimization that is needed to to add an object to manage
brainActivity. Summing in the collection of collections is taking
about half the current run time.  Denormalizing the data would help
significantly to make the code more readable, and performant then
using the collection of collections data storage.

Brain run will probably take a number of days to process.  This is
much improved over the week or more it would have taken before the
last optimization.

Currently the Brain Run does nothing. In smalltalk the system retains
access to the global variables after the system finishes running.
In python it is all gone.  We need to plum in the result presentations
that are needed. This will including determining the usefullness of the
X, XX, Y, YY, Z, ZZ global variables.  For now we will plug in the
present results workspace in the next commit.
(In other words there is no purpose in running the code to finish
because you will get no results when it is done except to find out
how long it took)

RJT